isWorking jsp sql

+++ src/main/Java/ru/user/dao/UserDAO.java
@@ -8,6 +8,6 @@ public interface UserDAO {
 	void add(User user);
 	void delete(User user);
 	void edit(User user);
-	User getById(int id);
+	User getById(Long id);
 	List<Role> allRoles();
 }
